# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão de Clientes",
        "description": "API para gestão e administração de clientes",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Julia Dourado",
            "url": "",
            "email": ""
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/projects": {
            "description": "Endpoints relacionados ao projeto",
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Stores a new project in the database",
                "operationId": "c6bde6b829987230e0e4d356ce57f863",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/admin/projects/{projectId}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Show project details",
                "description": "Displays details of the specified project",
                "operationId": "e6e367c10186b19ab8293ef7966880a9",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of the project to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "description": "Updates a project in the repository",
                "operationId": "5b7c208daf2c1bc5e695e71fefb81602",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of the project to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated successfully"
                    },
                    "404": {
                        "description": "Project not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "description": "Deletes a project from the repository",
                "operationId": "944d56a2237a5728c4dd74cf78a9e9de",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID of the project to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "required": [
                    "cpfcnpj",
                    "name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the client",
                        "type": "integer",
                        "example": 1
                    },
                    "cpfcnpj": {
                        "description": "The CPF or CNPJ of the client",
                        "type": "string",
                        "example": "12345678901"
                    },
                    "name": {
                        "description": "The name of the client",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "The email address of the client",
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "phone": {
                        "description": "The phone number of the client",
                        "type": "string",
                        "example": "+123456789"
                    },
                    "created_at": {
                        "description": "The date when the client was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "The date when the client was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-10T00:00:00Z"
                    },
                    "projects": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    }
                },
                "type": "object"
            },
            "Project": {
                "required": [
                    "name",
                    "description",
                    "client_id",
                    "location_uf",
                    "installation_type",
                    "equipment"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the project",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The name of the project",
                        "type": "string",
                        "example": "Project Alpha"
                    },
                    "description": {
                        "description": "A description of the project",
                        "type": "string",
                        "example": "This project involves the installation of solar panels."
                    },
                    "client_id": {
                        "description": "The ID of the client associated with the project",
                        "type": "integer",
                        "example": 1
                    },
                    "location_uf": {
                        "description": "The state (UF) where the project is located",
                        "type": "string",
                        "example": "SP"
                    },
                    "installation_type": {
                        "description": "The type of installation for the project",
                        "type": "string",
                        "example": "Solar Panel"
                    },
                    "equipment": {
                        "description": "The list of equipment used in the project",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Solar Panels"
                        }
                    },
                    "created_at": {
                        "description": "The date when the project was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "The date when the project was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-10T00:00:00Z"
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    }
                },
                "type": "object"
            },
            "StoreProjectRequest": {
                "required": [
                    "name",
                    "description",
                    "client_id",
                    "installation_type",
                    "location_uf",
                    "equipment"
                ],
                "properties": {
                    "name": {
                        "description": "Nome do projeto",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição do projeto",
                        "type": "string"
                    },
                    "client_id": {
                        "description": "ID do cliente",
                        "type": "integer"
                    },
                    "installation_type": {
                        "description": "Tipo de instalação",
                        "type": "string"
                    },
                    "location_uf": {
                        "description": "Estado de localização",
                        "type": "string"
                    },
                    "equipment": {
                        "description": "Equipamento usado",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateProjectRequest": {
                "required": [
                    "name",
                    "description",
                    "installation_type",
                    "location_uf",
                    "equipment"
                ],
                "properties": {
                    "name": {
                        "description": "Nome do projeto",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição do projeto",
                        "type": "string"
                    },
                    "installation_type": {
                        "description": "Tipo de instalação",
                        "type": "string"
                    },
                    "location_uf": {
                        "description": "Estado de localização",
                        "type": "string"
                    },
                    "equipment": {
                        "description": "Equipamento usado",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}